extract calculation into a service
